pyproject.toml
[project]
name = "google-apac-1"
version = "0.1.0"
description = ""
authors = [
    {name = "Manideep SP",email = "manideepsp.16@gmail.com"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "google-genai (>=1.70.0,<2.0.0)",
    "langchain-google-genai (>=4.2.1,<5.0.0)",
    "grpcio (>=1.80.0,<2.0.0)",
    "grpcio-tools (>=1.80.0,<2.0.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "google-auth (>=2.49.1,<3.0.0)",
    "google-api-python-client (>=2.193.0,<3.0.0)",
    "tavily-python (>=0.7.23,<0.8.0)",
    "mcp (>=1.27.0,<2.0.0)",
    "google-adk (>=1.28.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "google-apac-1"
version = "0.1.0"
description = ""
authors = ["your-name"]

package-mode = false